[CLN]: split where parsing into new file
